Refactoring of Tools
This is make tools applicable to ModelEntry in addition to WorkspaceEntry.

+++ PcompPlugin/src/org/workcraft/plugins/pcomp/tools/PcompTool.java
@@ -17,9 +17,11 @@ import org.workcraft.plugins.pcomp.tasks.PcompTask;
 import org.workcraft.plugins.stg.StgModel;
 import org.workcraft.plugins.stg.StgUtils;
 import org.workcraft.serialisation.Format;
+import org.workcraft.tasks.TaskManager;
 import org.workcraft.util.Export;
 import org.workcraft.util.FileUtils;
 import org.workcraft.util.GUI;
+import org.workcraft.util.LogUtils;
 import org.workcraft.util.WorkspaceUtils;
 import org.workcraft.workspace.ModelEntry;
 import org.workcraft.workspace.Workspace;
@@ -31,8 +33,8 @@ public class PcompTool implements Tool {
         return "Composition";
     }
 
-    public final boolean isApplicableTo(WorkspaceEntry we) {
-        return WorkspaceUtils.isApplicable(we, StgModel.class);
+    public final boolean isApplicableTo(ModelEntry me) {
+        return WorkspaceUtils.isApplicable(me, StgModel.class);
     }
 
     @Override
@@ -40,30 +42,44 @@ public class PcompTool implements Tool {
         return "Parallel composition [PComp]";
     }
 
-    public final void run(WorkspaceEntry we) {
+    @Override
+    public final ModelEntry run(ModelEntry me) {
         final Framework framework = Framework.getInstance();
-        MainWindow mainWindow = framework.getMainWindow();
-        PcompDialog dialog = new PcompDialog(mainWindow);
-        GUI.centerAndSizeToParent(dialog, mainWindow);
-        if (dialog.run()) {
-            String tmpPrefix = FileUtils.getTempPrefix("pcomp");
-            File tmpDirectory = FileUtils.createTempDirectory(tmpPrefix);
-            ArrayList<File> inputFiles = new ArrayList<>();
-            for (Path<String> path : dialog.getSourcePaths()) {
-                Workspace workspace = framework.getWorkspace();
-                File stgFile = exportStg(workspace.getOpenFile(path), tmpDirectory);
-                inputFiles.add(stgFile);
+        if (!framework.isInGuiMode()) {
+            LogUtils.logErrorLine("This tool only works in GUI mode.");
+        } else {
+            MainWindow mainWindow = framework.getMainWindow();
+            PcompDialog dialog = new PcompDialog(mainWindow);
+            GUI.centerAndSizeToParent(dialog, mainWindow);
+            if (dialog.run()) {
+                String tmpPrefix = FileUtils.getTempPrefix("pcomp");
+                File tmpDirectory = FileUtils.createTempDirectory(tmpPrefix);
+                ArrayList<File> inputFiles = new ArrayList<>();
+                for (Path<String> path : dialog.getSourcePaths()) {
+                    Workspace workspace = framework.getWorkspace();
+                    File stgFile = exportStg(workspace.getOpenFile(path), tmpDirectory);
+                    inputFiles.add(stgFile);
+                }
+
+                File outputFile = new File(tmpDirectory, "result.g");
+                outputFile.deleteOnExit();
+
+                PcompTask pcompTask = new PcompTask(inputFiles.toArray(new File[0]), outputFile, null,
+                        dialog.getMode(), dialog.isSharedOutputsChecked(), dialog.isImprovedPcompChecked(),
+                        tmpDirectory);
+
+                PcompResultHandler pcompResult = new PcompResultHandler(dialog.showInEditor(), outputFile);
+                TaskManager taskManager = framework.getTaskManager();
+                taskManager.queue(pcompTask, "Running parallel composition [PComp]", pcompResult);
             }
-
-            File outputFile = new File(tmpDirectory, "result.g");
-            outputFile.deleteOnExit();
-
-            PcompTask pcompTask = new PcompTask(inputFiles.toArray(new File[0]), outputFile, null,
-                    dialog.getMode(), dialog.isSharedOutputsChecked(), dialog.isImprovedPcompChecked(), tmpDirectory);
-
-            PcompResultHandler pcompResult = new PcompResultHandler(dialog.showInEditor(), outputFile);
-            framework.getTaskManager().queue(pcompTask, "Running parallel composition [PComp]", pcompResult);
         }
+        return me;
+    }
+
+    @Override
+    public WorkspaceEntry run(WorkspaceEntry we) {
+        run(we.getModelEntry());
+        return we;
     }
 
     public File exportStg(WorkspaceEntry we, File directory) {
